requirements.txt
slack-sdk
slack_bolt
hugchat
fastapi
uvicorn
python-dotenv
sentence-transformers 
scipy
weaviate-client